Refresh the full focused layout path on update
Instead of just checking focus on the most recent leaf widget. This
allows the layout to adjust focus to unexpected changes, such as
the saveslot dialog attempting to set the layout root to its submodal
before that modal's entrance transition has completed. Now, focus is
attached as soon as the submodal's widgets are available.